requirements.txt
streamlit==1.29.0
pandas==2.1.4
camelot-py==0.11.0
PyPDF2==3.0.1
opencv-python==4.8.1.78
ghostscript==0.7
pdf2image==1.16.3